// Angular/companies/companies-table-column-names.enum.ts
export enum CompaniesTableColumnNames {
  companyName = 'companyName',
  vat = 'vat',
  parentCompanyName = 'parentCompanyName',
  contactPersons = 'contactPersons',
  siteCountry = 'siteCountry',
  siteCity = 'siteCity',
  siteAddress = 'siteAddress',
  sitePhone = 'sitePhone',
  actions = 'actions'
}